+++ contracts/interfaces/IDegenerusGame.sol
@@ -272,7 +272,7 @@ interface IDegenerusGame {
     /// @param lootboxIndex Lootbox RNG index assigned at purchase time.
     /// @return amount Lootbox value in wei.
     /// @return presale True if presale mode is currently active (global flag, not per-lootbox).
-    function lootboxStatus(address player, uint32 lootboxIndex) external view returns (uint256 amount, bool presale);
+    function lootboxStatus(address player, uint48 lootboxIndex) external view returns (uint256 amount, bool presale);
 
     /// @notice Check whether lootbox presale mode is currently active.
     /// @return active True if presale is active.
@@ -282,7 +282,7 @@ interface IDegenerusGame {
     /// @dev Claims ETH, DGNRS, WWXRP, and potential boons/boosts.
     /// @param player The player address to open for (address(0) = msg.sender).
     /// @param lootboxIndex Lootbox RNG index assigned at purchase time.
-    function openLootBox(address player, uint32 lootboxIndex) external;
+    function openLootBox(address player, uint48 lootboxIndex) external;
 
     /// @notice Place Full Ticket Degenerette bets (4 traits, match-based payouts).
     /// @param player The betting player (address(0) = msg.sender).
